board 수정내용입니다.

+++ BUBBLEBEE/src/main/java/com/kh/bubblebee/board/model/vo/Board.java
@@ -12,8 +12,8 @@ public class Board {
 	private String ftitle; //제목
 	private String category; //카테고리
 	private String location; //장소
-	private int lat; //위도
-	private int lon; //경도
+	private double lat; //위도
+	private double lon; //경도
 	private String fcontain; //포함사항
 	private String fncontain; //불포함사항
 	private String fcalendar; //세부일정
@@ -36,7 +36,7 @@ public class Board {
 	public Board() {}
 
 	public Board(int rnum, int fno, int ftype, int price, String small_title, String ftitle, String category,
-			String location, int lat, int lon, String fcontain, String fncontain, String fcalendar, String fmaterials,
+			String location, double lat, double lon, String fcontain, String fncontain, String fcalendar, String fmaterials,
 			String fminfo, String fprecaution, String user_id, int heart, int rsum, int rcount, Date f_create_date,
 			String f_status, String originalFileName, String renameFileName, String introduce, int maxMember,
 			double rank, String hid) {
@@ -135,19 +135,19 @@ public class Board {
 		this.location = location;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLon() {
+	public double getLon() {
 		return lon;
 	}
 
-	public void setLon(int lon) {
+	public void setLon(double lon) {
 		this.lon = lon;
 	}
 
@@ -306,9 +306,4 @@ public class Board {
 				+ originalFileName + ", renameFileName=" + renameFileName + ", introduce=" + introduce + ", maxMember="
 				+ maxMember + ", rank=" + rank + ", hid=" + hid + "]";
 	}
-	
-	
-
-	
-	
 }
